rename components from login to auth, add links for navigation, replace formState
