+++ src/components/AuthPage/AuthPage.style.js
