+++ src/components/AuthPage/index.js
@@ -1,0 +1,3 @@
+import AuthPage from './AuthPage';
+
+export default AuthPage;
