--- src/components/LoginPage/index.js
@@ -1,3 +1,0 @@
-import LoginPage from "./LoginPage";
-
-export default LoginPage;
